move predicate byte handling to predicate pkg
