+++ plugin/evm/block.go
@@ -388,7 +388,7 @@ func (b *Block) verifyPredicates(predicateContext *precompileconfig.PredicateCon
 		return fmt.Errorf("failed to marshal predicate results: %w", err)
 	}
 	extraData := b.ethBlock.Extra()
-	headerPredicateResultsBytes := params.GetPredicateResultBytes(extraData)
+	headerPredicateResultsBytes := predicate.GetPredicateResultBytes(extraData)
 	if len(headerPredicateResultsBytes) == 0 {
 		return fmt.Errorf("failed to find predicate results in extra data: %x", extraData)
 	}
