+++ predicate/predicate_bytes.go
@@ -9,6 +9,10 @@ import (
 	"github.com/ethereum/go-ethereum/common"
 )
 
+// DynamicFeeExtraDataSize is defined in the predicate package to avoid a circular dependency.
+// After Durango, the extra data past the dynamic fee rollup window represents predicate results.
+const DynamicFeeExtraDataSize = 80
+
 // EndByte is used as a delimiter for the bytes packed into a precompile predicate.
 // Precompile predicates are encoded in the Access List of transactions in the access tuples
 // which means that its length must be a multiple of 32 (common.HashLength).
@@ -48,4 +52,25 @@ func UnpackPredicate(paddedPredicate []byte) ([]byte, error) {
 	}
 
 	return trimmedPredicateBytes[:len(trimmedPredicateBytes)-1], nil
+}
+
+// GetPredicateResultBytes returns the predicate result bytes from extraData. If
+// extraData is too short to include predicate results, it returns nil.
+func GetPredicateResultBytes(extraData []byte) []byte {
+	// Prior to Durango, the VM enforces the extra data is smaller than or equal
+	// to this size.
+	if len(extraData) <= DynamicFeeExtraDataSize {
+		return nil
+	}
+	// After Durango, the extra data past the dynamic fee rollup window represents
+	// predicate results.
+	return extraData[DynamicFeeExtraDataSize:]
+}
+
+// SetPredicateResultBytes sets the predicate results in the extraData in the
+// block header. This is used to set the predicate results in a block header
+// without modifying the initial portion of the extra data (dynamic fee window
+// rollup).
+func SetPredicateResultBytes(extraData []byte, predicateResults []byte) []byte {
+	return append(extraData[:DynamicFeeExtraDataSize], predicateResults...)
 }
