+++ predicate/predicate_bytes_test.go
@@ -48,3 +48,17 @@ func TestUnpackInvalidPredicate(t *testing.T) {
 		}
 	}
 }
+
+func TestPredicateResultsBytes(t *testing.T) {
+	require := require.New(t)
+	dataTooShort := utils.RandomBytes(DynamicFeeExtraDataSize - 1)
+	resultBytes := GetPredicateResultBytes(dataTooShort)
+	require.Empty(resultBytes)
+
+	preDurangoData := utils.RandomBytes(DynamicFeeExtraDataSize)
+	resultBytes = GetPredicateResultBytes(preDurangoData)
+	require.Empty(resultBytes)
+	postDurangoData := utils.RandomBytes(DynamicFeeExtraDataSize + 2)
+	resultBytes = GetPredicateResultBytes(postDurangoData)
+	require.Equal(resultBytes, postDurangoData[DynamicFeeExtraDataSize:])
+}
